requirements.txt
asgiref==3.8.1
Django==5.2.2
django-cors-headers==4.7.0
numpy==2.3.0
opencv-python==4.11.0.86
pillow==11.2.1
sqlparse==0.5.3
